Renamed `rules` package to `rule`
