+++ src/main/java/com/github/simonharmonicminor/beholder/core/configuration/rule/method/ProcedureRule.java
@@ -1,4 +1,4 @@
-package com.github.simonharmonicminor.beholder.core.configuration.rules.method;
+package com.github.simonharmonicminor.beholder.core.configuration.rule.method;
 
 /**
  * Represents a rule that defines restrictions to procedures
